CI: Skip uploading to Azure if no credentials
Upstream switched from S3 to Azure Storage for uploading binary blobs
during CI. (Used for serving releases from the official website?)

Just like we used to have it set up to skip uploading to Amazon S3
(due to this fork's lack of credentials), this commit adjusts
to the changes from upstream to also skip uploading to Azure.

(This fork doesn't have credentials for an Azure Storage account,
so skipping the upload is needed in order for CI to pass.)

(This first commit only adjusts the JS script that does the uploading.
Another change is needed to pass the env vars properly to the script.)