CI: Fix env vars passed to upload-artifacts script
Update the release-branch-build and nightly-release pipeleine YAML to:

- Add the _VAR_NAME_HERE intermediate var for Azure
- Delete unused Amazon S3-related vars
- Restore use of the intermediate vars (These make sure that vars are
  properly seen as empty/undefined in scripts, if unset in CI. They
  were accidentally reverted during a merge of upstream commits, I
  think?)

This makes sure the upload-artifacts.js script is able to properly
detect whether the env vars are unset, so it can skip attempting to
upload stuff to services we can't auth to (due to no credentials).

Skipping these uploads avoids errors, and thus allows CI to pass.