apm: Update to v2.6.5-atomic.1.1
Includes an updated Electron headers dist URL, as the atom.io one
is now down as a result of the Atom sunset.

Should allow apm to build native C/C++ packages again.

NOTE:
Fetching apm from its GitHub repo directly, instead of from the npm
package registry as before, as I do not have access to publish the
'@atom-ide-community/atom-package-manager' package to the npm package
registry.

This is the next best option. The commit this is pointing to has been
prepared equivalently to the process for uploading to the npm package
registry, by running `npm install` and `npm pack`, extracting the
resulting tarball, and committing (only) the result of those
operations.

Essentially pre-compiles the .coffee src/ files into .js lib/ files,
and prunes out a lot of un-needed files for Atom's purposes, such as
all the specs. Again, this is equivalent to how apm would be packaged
if I had been able to upload it to the npm package registry.

NOTE:
The commit this points to at the 'atom-community/apm' repo is the
commit at the tip of the '2.6.5-atomic.1.1-dist' branch at that repo,
as of the time this commit is being authored (28th January 2023).

NOTE:
Using a full git SHA so that this is content-addressable, and the file
content cannot change. Either this SHA is available at the git repo,
and the exact same file content can be downloaded from there, or the
`npm install` will fail. (Failing to install would be better than
surreptitiously downloading/running different/unexpected code.)

This is a best practice, in my opinion, when specifying dependencies
from a remote git repo, as the lockfile 'integrity' field is easier to
overlook when it changes, whereas this needs the exact SHA available,
or it will outright fail. Branches and tags are *not* immutable. SHAs
*are* immutable (barring some rather exotic hash collision attack).