docs: Update URL for the CI (binaries) dashboard
This makes the first half of the Installation docs followable again.

The old link went down, and while a redirect would be possible,
the atom-community org's forked version of the repo has deployed a fix
that isn't deployed at the original, and besides, we may as well bring
this resource into being officially hosted by the atom-community org,
rather than have it be maintained/hosted primarily on a personal
account, I suppose.